pyproject.toml
[project]
name = "maplord"
version = "2026.3.10.1"
description = "MapLord 2.0 — Real-time strategy web game on a world map"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-ninja>=1.5.3",
    "django-ninja-extra>=0.31.0",
    "django-ninja-jwt>=5.4.4",
    "channels>=4.2.0",
    "channels-redis>=4.2.1",
    "daphne>=4.1.0",
    "celery>=5.4.0",
    "redis>=5.2.0",
    "msgpack>=1.1.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "gunicorn>=23.0.0",
    "django-allauth>=65.0.0",
    "pillow>=11.0.0",
    "email-validator>=2.3.0",
    "django-cors-headers>=4.9.0",
]
